login: Process the case when install attributes fails to parse
We update the device enrollment status based on the data from install
attributes file. Failure to read that file should not trigger the
removal of enrollment, since that becomes an enrollment escape.

Also added log for the case when we explicitly remove enrollment to have
in case there's a scenario where it shouldn't happen.

BUG=chromium:1011375
TEST=Manually verified that both, consumer owned device and enrolled
device work as expected.

Change-Id: I1beed4795bfb834804cd6e1dd6ace6f630c15240
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1910229
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Commit-Queue: Igor <igorcov@chromium.org>