chromeos-config: match ARM architecture more broadly
veyron_mighty uses the architecture string "armv7l", which we did not
account for in our list of machine types. More broadly match arm
machine types by looking for "arm" or "aarch64" as the beginning of
the string.

The reason this matters is veyron_mighty, despite being non-unibuild,
needs to fallback to mosys, and architecture detection needs to happen
before mosys can be run.

BUG=chromium:991653,chromium:1009475
TEST=sudo FEATURES=test emerge chromeos-config-tools
TEST=cros_config /firmware image-name now runs "mosys platform model"
on veyron_mighty instead of failing

Change-Id: Iefa7f335f117b89894d6215bd0de75bd49a8d767
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1849427
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>