crash: Touch a file when anomaly detector is ready
Using logs to detect if the anomaly detector is ready (as we did before)
is unreliable and prone to failure if another system component has
overly verbose logging.

While logspam is a problem for the system, this test shouldn't fail
because of it, as it is unlikely that anyone would connect the two.

BUG=chromium:983725
TEST=tast -verbose run 100.127.92.23 platform.KernelWarning

Cq-Depend: chromium:1857465
Change-Id: Ic31a4646196d05ea19116a39dc6ce62b8a14d336
Reviewed-on: https://chromium-review.googlesource.com/1856681
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Ready: Miriam Zimmerman <mutexlox@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>