shill: pppoe: Use asynchronous Disconnect
Prior to this change, PPPoEService would directly destruct the pppd
ExternalTask. In the case that the Disconnect was not triggered by
OnPPPDied, this would cause PPPoEService to reach an idle state prior
to pppd terminating. If a Connect was called on that PPPoEService
right after the disconnection completed, this would lead to a race in
which a new pppd proccess would be started prior to the old one
terminating.

This change causes PPPoEService to remain in the disconnecting state
until pppd has fully terminated, preventing this race.

BUG=b:136288061, chromium:929781
TEST=All unit tests are passing.

Change-Id: I75cb873967a99b81e249f1e807fa051f6435ce70
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1857180
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>