biod: Do not emit error in log when checking EC command support
When checking whether a specific EC command is supported, we use
the EC command EC_CMD_GET_CMD_VERSIONS. If this command returns
with EC_RES_INVALID_PARAM, it only means the command to check is not
support, and doesn't mean anything failed. Logging an error in this
case would be confusing, so avoid emitting error. This does not
affect any functionality.

BUG=none
TEST=FEATURES=test emerge-nocturne biod
TEST=verified that the misleading error message no longer appears

Change-Id: I589124a9a1981268cf5bd23e79fbfc8b6394f862
Signed-off-by: Yicheng Li <yichengli@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1860915
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>