cros-disks: Add tests for SandboxedProcess
Increased the test coverage for SandboxedProcess.

A process can be run either with:
- Process.Run(&output)
- Process.Start() + Process.Wait()

Also, a process can be run with or without a PID namespace.

Each combination has its own code path, and we need to test them all.
Some bugs are only visible in one of the combinations.

BUG=chromium:1003654, chromium:1007613, chromium:1008226, chromium:1008637, chromium:1005642
TEST=cros_workon_make --board=eve chromeos-base/cros-disks --test

Change-Id: I099c9ab221ffd07de3e0d5c51d7adbcf1f045a14
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1865916
Reviewed-by: Sergei Datsenko <dats@chromium.org>
Commit-Queue: François Degros <fdegros@chromium.org>
Tested-by: François Degros <fdegros@chromium.org>