u2fd: Include allowlisting data when G2F enabled by policy
This policy will only ever be used in conjunction
with a server that expects to receive the allowlisting
data; we should therefore always include it.

We do not include the allowlisting data when G2F mode
is enabled from crosh, as there is anecdotal evidence
that some non-enterprise users are using G2F mode, and
would be broken if we included the allowlisting data.

BUG=b:135412897
TEST=enable g2f with policy, check allowlisting data is included

Change-Id: I08f6d50d397cd34da4b17ed3c55ac0f70a075545
Reviewed-on: https://chromium-review.googlesource.com/1869089
Tested-by: Louis Collard <louiscollard@chromium.org>
Commit-Ready: Louis Collard <louiscollard@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Wei-Cheng Xiao <garryxiao@chromium.org>