cryptohome: Allow mocking Process objects in tests.
cryptohome::Mount is already using the process class to start the file
attribute cleaner. This was printing errors in unit tests, though it
wasn't failing the tests because the cleaner is started as best-effort.

Use the Platform class to obtain Process objects so that they can be
mocked in tests. This is in preparation for moving cryptohome mounts
out-of-process, but has the immediate benefit of removing
file_attrs_cleaner_tool errors in unit tests.

BUG=chromium:985492
TEST=Unit tests pass.
TEST=WithEcryptfs/MountTest.MountCryptohomeShadowOnly/0
TEST=shows no process errors.

Change-Id: Ie6484e36f2b0059382833dffd3b1f3efaa685c5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1869300
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Sean Abraham <seanabraham@chromium.org>