crosh: capture_utility: update 50-crosh.sh to support VHT
Added support for VHT. This update is required for the script's updates
to function when invoked in crosh.

Additionally, added an error message explaining the "--output-file"
option. Previously, the shell script would print usage information
containing this flag, but when the flag was passed in crosh the program
would throw an "unknown option" error. Now, the program explains that
crosh is in charge of randomizing an output file in downloads.

BUG=chromium:1015108
TEST=`emerge-${BOARD} chromeos-base/crosh` &&
`cros deploy ${DUT_IP} chromeos-base/crosh`
Verify that VHT packets are being captured.

Cq-Depend: chromium:1872294
Change-Id: Iac7f4d2dae86e00240e1f6cd96298e83310b2ace
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1873385
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Jared Pauletti <pauletti@google.com>