camera: Workaround for Huddly GO deadlock.
This CL implements a similar workaround as
https://crrev.com/c/1859901 but for the new camera HAL
(cros_camera_service): when waiting for a captured frame from a
camera results in a timeout, we restart the stream. This gets
around the deadlock condition for the Huddly GO camera.

BUG=1010557
BUG=b:141978989
TEST=Visit
https://webrtc.github.io/samples/src/content/devices/input-output/
and continuously switch between the regular and overview streams
of a Huddly GO camera. We recover successfully from a timeout.

Change-Id: Ibc9f3aad82ab1f101c51bc601eca2e03285a4b76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1874596
Tested-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Andres Calderon Jaramillo <andrescj@google.com>
Reviewed-by: Shik Chen <shik@chromium.org>