hammerd: Use chromeos-config to query the model configuration
To query the different model configuration at runtime, we use
cros_config to query the target board name and its configuration.
To maintain backwards compatibility with the old style, we
check the ec-image-name, if we can't find it, then just fall
back to the old style.

BUG=b:139038006
TEST=checked the hammerd.log

Change-Id: Ie0e98ccedd23f066c159df6d9ae36496037b5ef1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1873847
Commit-Queue: Zhuohao Lee <zhuohao@chromium.org>
Tested-by: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>