cryptohome: create a util func to generate a secure random blob
The new util func is a wrapper that calls CryptoLib::GetSecureRandom() to
create a secure random blob with a given size.

In cryptohome, replace the repeating code pattern of creating a SecureBlob
object, setting its size, and getting random data for it with the new wrapper.

Also, add a TODO for handling corner cases in GetSecureRandom(). Adding error
handling will affect many places in cryptohome (see b:143445674), and I'd like
to create a bug to for that and have that done in other CLs.

BUG=b:142769655
TEST=unit tests;
     cros deploy cryptohome to a soraka and a chell, check cryptohomed syslog
     for both oobe and normal boots;
     PreCQ.

Change-Id: I32846c426a5622f37fc4812014a340b06c8feb93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1883153
Tested-by: Wei-Cheng Xiao <garryxiao@chromium.org>
Commit-Queue: Wei-Cheng Xiao <garryxiao@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>