camera: Fix field initialization order.
The next version of the compiler has a new warning, -Wreorder-init-list,
that triggers if struct field initialization assignments do not happen
in the same order in which the fields were declared.  This is causing
cros-camera-hal-mtk to fail to build for kukui.  This CL fixes that issue.

BUG=chromium:1017792
TEST=Tryjob with this CL succeeded:
https://ci.chromium.org/p/chromeos/builders/general/Try/b8898597060468056080

Change-Id: Id8354e6add7df28a2878e033e7b080b9ea63cf88
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1879488
Reviewed-by: MengHui Lin <menghui.lin@mediatek.com>
Reviewed-by: Shik Chen <shik@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Commit-Queue: Caroline Tice <cmtice@chromium.org>