vm_tools: Fix minor warnings for next version of compiler.
The next version of the compiler has a new warnings, -Wimplicit-fallthrough
(no 'break' for switch statement case) and -Wreorder-init-list,
(triggers if struct field initialization assignments do not happen
in the same order in which the fields were declared).  These warnings are
causing vm_guest_tools to fail to build for tael.  This CL fixes that issue.

BUG=chromium:1017817
TEST=Tael-full tryjob succeeded:
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1879490

Change-Id: I164f4c7f53a53d92f38240d5022677e9194044dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1879490
Tested-by: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: David Riley <davidriley@chromium.org>