arc: network: ndproxy: Use MAC address in neighbor entry if possible
When an ND message is proxied, if destination MAC address is a
unicast address, it needs to be replaced so that the destination
L3 stack can receive the packet correctly. In the previous version
we used broadcast MAC address in this situation. However, some
hosts (particular routers) apply a filter that will drop ND frames
with unicast destination IP and broadcast destination MAC.

In this patch, in order to address its issue, we look up in kernel
neighbor table first to find the unicast MAC address of
destination host and use that MAC for ND proxy. If the destination
IP cannot be found in neighbor table we fall back to the broadcast
MAC as a best-effort.

BUG=b:132574450
TEST=unit, manual

Change-Id: I6bfef92c0da3029f7e337d97e85d880716d919ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1885539
Tested-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Auto-Submit: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Taoyu Li <taoyl@chromium.org>