cryptohome: Use a clean-up closure.
A previous CL introduced a ScopedClosureRunner to Mount* functions to
ensure that clean-up could be handled consistently across code paths.
However, the UnmountCryptohome function was not benefitting from this
new logic. For example, it still attempts to call CleanUpEphemeral()
even when the mount was not an ephemeral mount.

Improve this by centralizing clean-up and unmounting code in a closure
and storing this closure when a mount operation succeeds, and then
using it in UnmountCryptohome().

This will also allow piecemeal moving to out-of-process mounting
without modifying the basic Mount/Unmount flow in cryptohome.
In this CL MountEphemeralCryptohome() receives the same clean-up
closure from all call sites, but the next CL after this will pass
a different clean-up closure for the Guest session case.

BUG=chromium:985492
TEST=-unit tests pass.
TEST=Deploy to device, log in as regular user, log in as guest.

Change-Id: I904f580449dddd56d103fc11d53505aed1a2815d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1881719
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>