metrics: Make CumulativeMetrics::Update public
While CumulativeMetrics::Update is called periodically, sometimes an
event occurs that warrants an earlier update. To accommodate this
situation, this change makes CumulativeMetrics::Update public. Since
the accumulated value is not a float/double, we don't have to worry
about more frequent updates accumulating rounding errors.

For example, CL:1880317 reworks the existing metrics for accumulating
the time over a certain period in which a particular technology (WiFi,
ethernet, etc) is used for the primary connection. Without triggering
a metric update when the default connection changes, the metrics will
accumulate inaccuracies, which could be non-negligible over a period
of a month. On the other hand, the periodic update is still useful for
resilience to abnormal terminations like crashes.

BUG=chromium:1004325
TEST=All unit tests are passing.

Change-Id: I86478e42d1a5ef6fa5cd68079ac7c14f441fc356
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1888300
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>