shill: device: Add AsWeakPtr method
Future lower Device changes will involve a VirtualDevice referring to
its lower Device if there is one. Since we don't want a VirtualDevice
to control the lifetime of its lower Device, it should either hold a
weak ptr or a raw ptr rather than shared ptr. Using a weak ptr gives
us better guarantees (we don't need shill to crash more
frequently). This change also moves internal uses over to AsWeakPtr
for convenience.

In addition, the WeakPtrFactory is moved to the bottom of Device (as
weak_ptr.h already says we should be doing) so that WeakPtrs are
invalidated at the beginning of Device destruction. Currently, we only
use Device weak_ptrs for task posting (which this doesn't affect) and
for the IPConfig callbacks. Given that IPConfig does not call any of
its callbacks during destruction, this should not result in a change
of behavior.

BUG=chromium:999589
TEST=All unit tests are passing.

Change-Id: If106a1f74d45040fcc3c62aecd014f1c7b6c21fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1949116
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>