shill: Support subject alternative name match.
Subject alternative name match is used by the WPA Supplicant to be
matched against the alternative subject name of the authentication
server certificate. It is sent from chrome to shill to WPA supplicant.
It is sent to shill in a serialized form to be able to store it
permanently on disk, because StoreInterface used in chromeOS to save EAP
data to permanently doesn't support saving dictionaries. It is
deserialized and then converted to the format needed by the WPA
supplicant, which is a semicolon separated string of entries.
Chrome implementation that uses this CL:1892955

BUG=chromium:1004261
TEST=shill unit tests
Change-Id: I41fcc096cf9c2f41ecb23f3efc993f27bb01c0ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1897535
Tested-by: Omar Morsi <omorsi@google.com>
Commit-Queue: Omar Morsi <omorsi@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>