cryptohome: Converge HandleOwnershipTakenSignal() and OwnershipCallback()
HandleOwnershipTakenSignal() is called when cryptohome is notified that the
TPM ownership is taken, and OwnershipCallback() is called when ownership is
taken due to action from cryptohome's side. This change merge the two callback
to ensure consistent behaviour regardless of the source of ownership taken
event.

BUG=none
TEST=manual && FEATURES=test USE=cryptohome_userdataauth_interface emerge-$BOARD cryptohome

Change-Id: Ic2fea8958e2772f6b04589416eb1e78de6886135
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1898227
Tested-by: John L Chen <zuan@chromium.org>
Commit-Queue: John L Chen <zuan@chromium.org>
Reviewed-by: Wei-Cheng Xiao <garryxiao@chromium.org>