hammerd: Fix asan errors in update_fw_test
The address sanitizer fails to run this test when the fmap isn't aligned
to a word and when it tries to read the mock fmap area's size in the
ASSERT_EQ() test. Pad out the fmap by 8 bytes so it's guaranteed to be
aligned and move the mock fmap size dereference outside of the
ASSERT_EQ() check so that asan doesn't complain about this test.

BUG=None
TEST=USE="asan" FEATURES="test" emerge-amd64-generic hammerd
Change-Id: I40497ba814d2292637dcc1ead76506b465101526
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1898034
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Auto-Submit: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Joel Kitching <kitching@chromium.org>