shill: connection: Use "priority" rather than "metric"
Linux uses the terminology "metric" specifically for routes, while
"priority" is used for routing rules. Let's move over for consistency
and clarity. Note that |priority_| is sometimes used as a metric
value. A comment above |priority_| is added to explain why this is
done.

BUG=None
TEST=All unit tests are passing.

Change-Id: Ia577a514433cc87e2915fca116a2a32073746e28
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1899103
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>