libbrillo: Fix fuzzer recursion bug in http_form_data_fuzzer
Add a stack depth limit when building the multipart form fields to
prevent crashing in the fuzzer on stack depth. Switching to iterative
isn't an ideal solution because the libbrillo code is recursive as well
and it does not seem worthwhile to make that iterative.

BUG=chromium:1017809
TEST=Test case no longer crashes

Change-Id: Id5b886390be36538ff38103addef643cc1fd8573
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1900316
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>