shill: remove synchronous error from Manager::SetEnabledStateForTechnology
The synchronous/asynchronous error conditions had to be combined
in the D-Bus adaptor anyway, so there is no benefit to exposing
them in lower APIs. Splitting the error reporting also caused the
method to reimplement much of ResultAggregator for the synchronous
case, which was needless.

BUG=chromium:1020264
TEST=unit tests

Change-Id: Ifc5f07b886cfe2aecf053e6fa9a1aa83367008c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1900499
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>