diagnostics: replace EcEventService::Delegate
Replace EcEventService::Delegate with Observer instead. This is to
provide better isolation of EcEventService into telemetry/, also:
 - WilcoDtcSupportd services now depend on the lower-level telemetry/
   services.
 - EcEventService doesn't need to know about the IPC technology used but
   rather is only responsible for delivering Ec events.

Move relevant tests from EcEventService to WilcoDtcSupportdCore
that test various Mojo events triggered by Ec events as well as gRPC
calls.

Additionally, fixing the bug in tests where the bootstrapping of the mojo
connection is done asynchronously without waiting the initialization
to complete. This issue wasn't obvious before because the bootstrapping
was done while making gRPC call. The gRPC call adds just enough time for
the mojo connection to be established.
However, while adding the test where we emit ec events and we
expect the browser to receive the mojo call, the mojo_service_ object
was null.
The fix is done by adding a callback to be called on completion, and
from the test side, we run the BootstrapMojoConnection in a run_loop.

BUG=b:143861345
TEST=cros_workon_make --board=sarien diagnostics --test

Change-Id: I3b300a872c2b44892a1a28c3d5310105c6849d60
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1899689
Tested-by: Mahmoud Gawad <mgawad@google.com>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Commit-Queue: Mahmoud Gawad <mgawad@google.com>