tpm_manager, cryptohome: move tpm version info out of tpm status
This is the first step of caching TPM version info and status in tpm_managerd.
Version info is unrelated to other info in tpm status, and moving it out of tpm
status will simplify the work of caching. Actual changes to caching will be in
upcoming CLs.

Also fixed a bug in legacy cryptohome proxy that non-UTF8 vendor specific data
cannot be converted to a string in the dbus response. Fixed by hex-encoding the
data, which is also consistent with what tpm_manager_client and
cryptohome::Service::TpmGetVersionStructured() do.

BUG=chromium:1016698
TEST=unit tests;
     * Tested on a soraka and chell:
        1. tpm_manager_client get_version_info
        2. stop cryptohomed, run cryptohome-proxy, then run
           dbus-send --system --print-reply --dest=org.chromium.Cryptohome \
                     /org/chromium/Cryptohome \
                     org.chromium.CryptohomeInterface.TpmGetVersionStructured
        1) and 2) successfully executed and returned the same results.
     * Tested `tpm-manager get_version_info` on a soraka.
     * Verified that GetTpmStatus commands/workflows still work.

Change-Id: I09dfe8d097a3426503cab2224ac1f50cc61ef484
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1911824
Tested-by: Wei-Cheng Xiao <garryxiao@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: John L Chen <zuan@chromium.org>
Commit-Queue: Wei-Cheng Xiao <garryxiao@chromium.org>