attestation: fix a failing unit test on cheza64
The test TpmUtilityTest.CreateCertifiedKeyWithEccCertified failed if run
with its previous test but passed if run alone. In CreateCertifiedKey(),
the uninitialized key_handle happens to be 0x12345678, which matches the 2nd
GetKeyPublicArea EXPECT_CALL rule and causes a wrong public key type to be
returned, which is unexpected by the test.

Explicitly specifying a different fake key handle and use it in mock LoadKey
and GetKeyPublicArea fixes the test.

BUG=chromium:1022241
TEST=FEATURES=test emerge-cheza64 attestation

Change-Id: I2293cb1b9e07ee322875382911f0a10aa309d263
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1902839
Tested-by: Wei-Cheng Xiao <garryxiao@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>