shill: Centralize recognition of "arcbr0" in DeviceInfo
Prior to this change, the init script contained the name "arcbr0",
which was then passed as a command-line argument and made its way
through shill_main -> ShillDaemon -> DaemonTask -> Manager ->
DeviceInfo. With this information, DeviceInfo could then recognize the
arcbr0 interface as the ARC bridge. This change centralizes the arcbr0
information in DeviceInfo, which generally contains the logic for
distinguishing different interfaces based on various properties
including link name.

Note that VPNProvider still creates the VirtualDevice representing
arcbr0. This will likely change in a future CL to have DeviceInfo
create the arcbr0 Device when we have a proper representation for
guest interfaces. For now, the arcbr0 VirtualDevice is still only used
for ARC VPNs, so it makes sense to leave it here until that changes.

BUG=chromium:1010302
TEST=All unit tests are passing.

Change-Id: Ia6ee8451f7b8c7e7cec8f57ef1283e2574e5bcc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1904249
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>