shill: Remove unused source address blackhole functionality
Prior to moving to per-Device routing, source address blackholing was
used as an attempt to work around the numerous routing issues on the
system. This code is no longer used, and it's unlikely that it will be
used in the future given that per-Device routing generally precludes
the need to blackhole traffic based on source address. This change
removes the source address blackhole code.

BUG=chromium:1022028
TEST=All unit tests are passing.

Change-Id: Ib312602b099f23505f59a8ff4a1f2a1e8209387a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1904251
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>