cryptohome: Fix size of AES-GCM keys and IVs.
The constants for the size of AES-GCM keys and IVs are incorrect because
they divide the number of bits by the sizeof(uint8_t), which returns a
byte value. sizeof(uint8_t) must be multipled by CHAR_BIT to get the
size of a byte in bits.

BUG=chromium:728125
TEST=FEATURES="test" emerge-${BOARD} cryptohome

Change-Id: I91ab16ecc5cfdc6daabca8711c7138cab584d488
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1907466
Tested-by: Greg Kerr <kerrnel@chromium.org>
Auto-Submit: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>
Commit-Queue: Greg Kerr <kerrnel@chromium.org>