chromeos-config: force UTF-8 encoding on readme tests
The tests pass when the locale encoding is UTF-8, but not when it's
ASCII.  We'd use encoding to the open() call, but Python 2 doesn't
support that, so use a manual decode hack.  When we drop Python 2,
we can clean this up and use open() encoding everywhere.

BUG=chromium:999879
TEST=tests pass w/python3 w/LC_ALL=C

Change-Id: I2c78a7db1b6d659f533832e5cd816dfd140761b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1907437
Reviewed-by: Ned Nguyen <nednguyen@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Ned Nguyen <nednguyen@google.com>
Tested-by: Ned Nguyen <nednguyen@google.com>