shill: test-scripts: Create list-manager script
This script is quite similar to the test-manager script that was
removed in CL:1429739, which allows users to see Manager
properties. One benefit of this implementation is that it uses the
convert_dbus_value script, which precludes the need for updating this
script to keep track of property types (one of the reasons
test-manager became obsolete).

BUG=None
TEST=Deploy and run `./list-manager`.

Change-Id: I6c864a1f87d3b74d2e82cd7def7911b708b7d053
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1910498
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>