shill: routing_table: Reset accept_ra_rt_table to 0 on DeregisterDevice
Prior to this change, RoutingTable did not reset an interface's
accept_ra_rt_table to 0 when deregistering it. The logic behind this
decision was that resetting accept_ra_rt_table to 0 does *not* move
all of the routes back to the main routing table. Furthermore, if the
entry in /proc/sys/net/ipv6/conf/ is removed and later recreated,
accept_ra_rt_table will be back to the default of 0 anyways. However,
in some cases an (e.g., bridge) interface will be created only
once. If shill were to switch between managing and not managing such
an interface, it would be necessary for shill to reset the
accept_ra_rt_table value in this case. Hence this change causes
RoutingTable to reset the accept_ra_rt_table value when deregistering
a Device.

In addition, RoutingTable::DeregisterDevice was previously not being
triggered by DeviceInfo. This change rectifies that oversight.

BUG=chromium:941597, b:142245073
TEST=All unit tests are passing.
TEST=Manual testing to ensure that removing and adding an interface
     still results in accept_ra_rt_table being set to -1000, routes being
     where they should be, and a "Device $N deregistered." log from
     RoutingTable.

Change-Id: I3ae1437187d3ba8a3b7e6dcfa1db1d9b55446635
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1913638
Tested-by: Alex Khouderchah <akhouderchah@chromium.org>
Commit-Queue: Alex Khouderchah <akhouderchah@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>