dlcservice: Re-architect DBus management
Currently All DBus and common service functionalities are shoved into
one class and its hard to change or test it properly. This CL breaks the
DlcServiceDbusAdaptor into three classes:
- DlcService: which has the core functionality of dlcservice.
- DBusService: Which is a wrapper around DlcService functions to
  provide public DBus functions.
- DBusAdaptor: Is the DBus adaptor class we need to send back stuff like
  signals, etc.

This CL also changes the ordering of returned Errors as return values
should be the last arguments in the class, not the beginning.

Also there is no need to forward declate the BootSlot.

BUG=chromium:1019470
TEST=FEATURES=test emerge dlcservice-client dlcservice

Change-Id: I52c68f3e4f76b8bb3aab534866925ae822de215b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1918693
Tested-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>