vm_tools: concierge: fix disconnecting CDROMs after creating Plugin VM
The helper tool does not return "connected" status for CD-ROMS that are
connected, but rather only return "disconnected" status when devices are
disconnected. Change the logic to skip disconnect if status is present
in the device's properties and its value matches "disconnected".

BUG=b:144526327
TEST=Create a new VM, verify tools CD is disconnected after tools
install is complete.

Change-Id: I26c64b04df84286e2f9bbeb6d1e9a173d682df9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2037901
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>