cros-disks: Change Uri::Parse to be more robust
Modified Uri::Parse to not require as a precondition that the passed
string be a valid URI. Instead, if the passed string cannot be parsed,
then Uri::Parse returns an invalid Uri object with an empty scheme which
can be tested. This avoids calling Uri::IsUri before calling Uri::Parse,
which used to parse the string twice. Now, it is parsed only once, and
the logic is not duplicated anymore.

BUG=chromium:996549
TEST=cros_workon_make --board=eve chromeos-base/cros-disks --test

Change-Id: Ie78cfb51bcfa1df77e3f45231408cf81ab171797
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1923708
Tested-by: François Degros <fdegros@chromium.org>
Reviewed-by: Sergei Datsenko <dats@chromium.org>