vm_tools: concierge: allow creation of preallocated disks
If the `disk_size` parameter is specified (non-zero) in the
CreateDiskImageRequest, treat the disk image as having a user-chosen
size.

A user-chosen-size disk image will have two properties that distinguish it
from the current default of sparse disk images:
- The disk will not be automatically grown to 90% of free space.
- The space for the disk image will be preallocated.

In order to accomplish the first point, set the fixed_size xattr
introduced in CL:1904827 to prevent automatic resizing.

For the second point, the new codepath uses fallocate() to preallocate
the image instead of ftruncate() to leave it sparse.

This change will not affect the normal VM creation flow as it works
today, since both Chrome and `vmc start` pass 0 for `disk_size` in
CreateDiskImageRequest.  The new behavior will only be triggered once
the setup flow passes a user-chosen disk size.

This can be tested using the concierge_client command in dev mode:

  concierge_client --create_disk --cryptohome_id=... \
    --disk_path=termina --disk_size=$((3 * 1024 * 1024 * 1024))
  # Both ls -l and du -bs should return 3 GiB.
  ls -l  /home/root/.../crosvm/dGVybWluYQ\=\=.img
  du -bs /home/root/.../crosvm/dGVybWluYQ\=\=.img

BUG=chromium:858815
TEST=concierge_client command (see above) on grunt

Change-Id: I95cc81d8cd0e29e146669e8edfec89b720eef3dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1949461
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>