arc: networkd: Add Crostini service
This patch adds an implementation of a guest service for handling
Termina VMs from Crostini. It replicates the address allocation that
happens currently in Concierge. We continue to use names to tell
DeviceManager how to make each device but this will be refined in later
patches.

BUG=b:133194715
TEST=units, manual on atlas - termina only & with arcvm

Change-Id: I76072f0ca7a976480c94b611a800de501d12eaf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1932514
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>