smbfs: Only allow Kerberos setup to be run once
In smbprovider, SetupKerberos() could be called more than once because
smbprovider was a system service that remained active across browser
restarts (ie. on crash or flags change). In smbfs, this is no longer
possible because the lifecycle of an smbfs process is tied to a specific
browser process due to using Mojo. Because of this, SetupKerberos() no
longer needs to handle being called multiple times and the class can be
simplified.

BUG=chromium:939235
TEST=cros_run_unit_tests --board=chell --packages smbfs

Change-Id: Id5cffd18d276d00dc16612b6df77ffcca95b20e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1928418
Reviewed-by: Sergei Datsenko <dats@chromium.org>
Tested-by: Anand Mistry <amistry@chromium.org>
Commit-Queue: Anand Mistry <amistry@chromium.org>