arc: networkd: Add a sticky bit device option.
This option replaces the use of names as the means by which the device
manager decides on which devices should be kept independent of the
updates from shill. In particular, right now, this is used for keeping
the android device around but it will also be used to tell the device
manager to ignore termina tap devices.

BUG=b:133194715
TEST=units

Change-Id: I5eab4bf00a22c0ff1bce81a6f4c62c9604688ed4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1932512
Tested-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>