crash: Add crash-loss measurement
Attempt to measure the number of times that Chrome crashes but does not
call crash_reporter. We do this by watching the logs for messages
indicating that crash_reporter was called by Chrome and for messages
indicating that the kernel noticed that Chrome crashed, and then
matching them up.

BUG=chromium:1021374
TEST=Crashed Chrome and then checked chrome://histograms. Enabled
crashpad and repeated. Set up Chrome so that it wouldn't call
crash_reporter and repeated.

Change-Id: Ia4fad0b8377c81588e5fef210af75967b4d2b38c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1940859
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>